fix compile errors with clang-15 (#20069)
* fix compile errors with clang-15

* fix clang-format

+++ lib/Greenspun/lib/DateTime.cpp
@@ -57,7 +57,7 @@ EvalResult DateTime_dateStringToUnix(Machine& ctx, VPackSlice const paramsList,
     return EvalError("string did not parse as date");
   }
 
-  result.add(VPackValue(tp.time_since_epoch().count()));
+  result.add(VPackValue(static_cast<uint64_t>(tp.time_since_epoch().count())));
   return {};
 }
